requirements.txt
anthropic>=0.34.0
flask>=3.0.0
python-dotenv>=1.0.0
twilio>=9.0.0
requests>=2.31.0
gunicorn>=21.0.0
